requirements.txt
pydantic
pymongo
requests
cassandra-driver
time_uuid
pydgraph